Fixes CR defects: Better naming/methods and type ordering/split actual options and context helpers
